Fix asy compilation on Ubuntu 20.04
On Ubuntu 20.04 (and previous?) asy fails to run in the sandbox with
"Wrong __data_start/_end pair". After the investigation it seems to be
caused by a bug in libgc (fixed in 8.0.4), triggered by the unlimited
stack the sandbox sets.

The workaround is to limit the stack size just for asy.

Co-Authored-By: Alessandro Bortolin <bortolin.alessandro@outlook.it>